feat(lib,demo): Add support for reflective Coq output
This is related to the IFirstOrderDeep.v development from earlier in the
summer. It has been greatly simplified, however.

Not all proofs can be currently produced in this mode, since it is still
considerably unclear if the freeform treatment of init/congr can be achieved in
a purely reflective style. After a lot of attempts, I think it doesn't make
sense to complicate the reflective library too much further. Instead, the
To.Coq_reflect module needs to output more hints for the init/congr rules.

To be investigated further in future work.