feat(lib,demo): use symbolic formatting in `To_katex`
This inserts fake "break" and "space" nodes into the stream, which are then
turned into actual breaks and spaces in JavaScript.

It has to be done this way because the `\newline` command of katex is not
working correctly when it occurs nested inside a bunch of enclosed
groups. Funnily enough, the JavaScript output ends up looking and performing
much better than if katex were to do it.